Fix ambiguous GMP conversion with explicit cast to long
- Use static_cast<long> to explicitly cast int64_t values
- Avoids ambiguous conversion on macOS where int64_t is long long
- GMP's mpq_class has unambiguous constructor for long type

+++ fracessa/include/rational_linalg/types_rational.hpp
@@ -144,9 +144,9 @@ inline rational small_to_rational(const small_rational& val) {
     // Extract underlying int64_t values from rational64
     int64_t num = val.numerator();
     int64_t den = val.denominator();
-    // Construct rational from numerator and denominator using two-parameter constructor
-    // This avoids ambiguous conversion on macOS where int64_t is long long
-    return rational(num, den);
+    // Construct rational from numerator and denominator
+    // Explicitly cast to long to avoid ambiguous conversion on macOS where int64_t is long long
+    return rational(static_cast<long>(num), static_cast<long>(den));
 }
